중심좌표 설정전

+++ src/main/java/kr/kro/bbanggil/bakery/dto/BakeryReviewDTO.java
@@ -12,7 +12,7 @@ public class BakeryReviewDTO {
 	private int orderNo;
 	private int userNo;
 	private String reviewDetail;
-	private int reviewRating;
+	private double reviewRating;
 	private String reviewDate;
 	private int bakeryNo;
 }
